Split GUI metadata tag tests
