+++ src/gui_app/tests.rs
@@ -16,6 +16,7 @@ use std::{collections::HashMap, fs, path::PathBuf, sync::mpsc, time::Duration};
 
 mod audio_settings_controls;
 mod audio_settings_dropdowns;
+mod metadata_tag_tests;
 mod native_file_drop;
 mod sample_browser;
 mod shortcuts_context;
@@ -1099,1687 +1100,6 @@ fn default_folder_browser_loads_assets_root() {
             .iter()
             .any(|file| file.name == "portal_SS_kick_003.wav")
     );
-}
-
-#[test]
-fn folder_browser_sidebar_paints_filter_and_metadata_sections() {
-    let browser = super::FolderBrowserState::load_default();
-    let tags = vec![String::from("kick")];
-    let frame = radiant::runtime::UiSurface::new(
-        super::folder_browser::folder_browser_view(
-            &browser,
-            260.0,
-            true,
-            "",
-            &[],
-            None,
-            "add tag",
-            None,
-            &[],
-            &tags,
-            &[],
-            None,
-        )
-        .into_node(),
-    )
-    .frame(
-        Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(260.0, 620.0)),
-        &radiant::theme::ThemeTokens::default(),
-    );
-
-    assert!(frame_has_text(&frame, "Filter"));
-    assert!(!frame_has_text(&frame, "Metadata"));
-    assert!(!frame_has_text(&frame, "Tagging"));
-    assert!(frame_has_text(&frame, "kick"));
-    assert!(frame_has_text(&frame, ">"));
-}
-
-#[test]
-fn folder_browser_metadata_selected_tag_chip_uses_strong_accent_style() {
-    let browser = super::FolderBrowserState::load_default();
-    let tags = vec![String::from("hat")];
-    let theme = radiant::theme::ThemeTokens::default();
-    let frame = radiant::runtime::UiSurface::new(
-        super::folder_browser::folder_browser_view(
-            &browser,
-            260.0,
-            true,
-            "",
-            &[],
-            None,
-            "add tag",
-            None,
-            &[],
-            &tags,
-            &[],
-            Some("hat"),
-        )
-        .into_node(),
-    )
-    .frame(
-        Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(260.0, 620.0)),
-        &theme,
-    );
-
-    let tag_text = frame
-        .paint_plan
-        .primitives
-        .iter()
-        .find_map(|primitive| match primitive {
-            PaintPrimitive::Text(text) if text.text.as_str() == "hat" => Some(text),
-            _ => None,
-        })
-        .expect("selected tag chip should paint");
-    assert_eq!(tag_text.color, theme.text_primary);
-}
-
-#[test]
-fn clicking_metadata_tag_chip_selects_it_in_sidebar() {
-    let (mut state, _source_root, selected_file) = gui_state_with_temp_sample("tag-target.wav");
-    state
-        .metadata_tags_by_file
-        .insert(selected_file, vec![String::from("hat")]);
-    let bridge = DeclarativeOwnedRuntimeBridge::new(
-        state,
-        |state| radiant::runtime::UiSurface::new(super::view(state).into_node()),
-        |state, message| state.apply_message(message, &mut ui::UpdateContext::default()),
-    );
-    let mut runtime = SurfaceRuntime::new(bridge, Vector2::new(900.0, 620.0));
-    let tag_rect = text_rect(
-        &runtime.frame(&radiant::theme::ThemeTokens::default()),
-        "hat",
-    )
-    .expect("metadata tag chip should paint");
-    let point = Point::new(
-        (tag_rect.min.x + tag_rect.max.x) * 0.5,
-        (tag_rect.min.y + tag_rect.max.y) * 0.5,
-    );
-
-    runtime.dispatch_event(Event::PointerPress {
-        position: point,
-        button: PointerButton::Primary,
-        modifiers: PointerModifiers::default(),
-    });
-    runtime.dispatch_message(super::GuiMessage::Frame);
-    runtime.dispatch_event(Event::PointerRelease {
-        position: point,
-        button: PointerButton::Primary,
-        modifiers: PointerModifiers::default(),
-    });
-
-    assert_eq!(
-        runtime.bridge().state().selected_metadata_tag.as_deref(),
-        Some("hat")
-    );
-}
-
-#[test]
-fn metadata_tag_chips_display_playback_tags_first() {
-    let (mut state, _source_root, selected_file) = gui_state_with_temp_sample("tag-target.wav");
-    state.metadata_tags_by_file.insert(
-        selected_file,
-        vec![
-            String::from("hat"),
-            String::from("warm"),
-            String::from("loop"),
-            String::from("one-shot"),
-        ],
-    );
-
-    let frame = radiant::runtime::UiSurface::new(super::view(&mut state).into_node()).frame(
-        Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(900.0, 620.0)),
-        &radiant::theme::ThemeTokens::default(),
-    );
-
-    let loop_rect = text_rect(&frame, "loop").expect("loop tag should paint");
-    let one_shot_rect = text_rect(&frame, "one-shot").expect("one-shot tag should paint");
-    let hat_rect = text_rect(&frame, "hat").expect("hat tag should paint");
-    let warm_rect = text_rect(&frame, "warm").expect("warm tag should paint");
-
-    assert!(loop_rect.min.x < hat_rect.min.x);
-    assert!(one_shot_rect.min.x < hat_rect.min.x);
-    assert!(hat_rect.min.x < warm_rect.min.x);
-}
-
-#[test]
-fn metadata_tag_chips_group_by_target_category_order_and_color() {
-    let browser = super::FolderBrowserState::load_default();
-    let tags = vec![
-        String::from("warm"),
-        String::from("artist1"),
-        String::from("dorian"),
-        String::from("hat"),
-        String::from("loop"),
-    ];
-    let categories = vec![
-        super::metadata_tags::MetadataTagDisplayCategory {
-            tag: String::from("warm"),
-            category_id: "character",
-        },
-        super::metadata_tags::MetadataTagDisplayCategory {
-            tag: String::from("artist1"),
-            category_id: "prefix",
-        },
-        super::metadata_tags::MetadataTagDisplayCategory {
-            tag: String::from("dorian"),
-            category_id: "tuning-scale",
-        },
-        super::metadata_tags::MetadataTagDisplayCategory {
-            tag: String::from("hat"),
-            category_id: "sound-type",
-        },
-        super::metadata_tags::MetadataTagDisplayCategory {
-            tag: String::from("loop"),
-            category_id: "playback-type",
-        },
-    ];
-    let theme = radiant::theme::ThemeTokens::default();
-    let frame = radiant::runtime::UiSurface::new(
-        super::folder_browser::folder_browser_view(
-            &browser,
-            600.0,
-            true,
-            "",
-            &[],
-            None,
-            "add tag",
-            None,
-            &[],
-            &tags,
-            categories.as_slice(),
-            None,
-        )
-        .into_node(),
-    )
-    .frame(
-        Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(600.0, 620.0)),
-        &theme,
-    );
-
-    let loop_rect = text_rect(&frame, "loop").expect("loop tag should paint");
-    let hat_rect = text_rect(&frame, "hat").expect("hat tag should paint");
-    let warm_rect = text_rect(&frame, "warm").expect("warm tag should paint");
-    let artist_rect = text_rect(&frame, "artist1").expect("prefix tag should paint");
-    let dorian_rect = text_rect(&frame, "dorian").expect("tuning tag should paint");
-
-    assert!(loop_rect.min.x < hat_rect.min.x);
-    assert!(hat_rect.min.x < warm_rect.min.x);
-    assert!(warm_rect.min.x < artist_rect.min.x);
-    assert!(artist_rect.min.x < dorian_rect.min.x);
-
-    assert_eq!(text_color(&frame, "loop"), Some(theme.bg_primary));
-    assert_eq!(text_color(&frame, "hat"), Some(theme.accent_mint));
-    assert_eq!(text_color(&frame, "warm"), Some(theme.highlight_cyan));
-    assert_eq!(text_color(&frame, "artist1"), Some(theme.accent_danger));
-    assert_eq!(text_color(&frame, "dorian"), Some(theme.text_muted));
-}
-
-#[test]
-fn default_gui_tag_library_opens_beside_folder_sidebar() {
-    let (mut state, _source_root, selected_file) = gui_state_with_temp_sample("tag-target.wav");
-    state.metadata_tags_by_file.insert(
-        selected_file.clone(),
-        vec![String::from("hat"), String::from("seq")],
-    );
-    state.metadata_tags_by_file.insert(
-        String::from("other.wav"),
-        vec![String::from("bass"), String::from("hat")],
-    );
-    state.metadata_tag_library_open = true;
-
-    let frame = radiant::runtime::UiSurface::new(super::view(&mut state).into_node()).frame(
-        Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(900.0, 620.0)),
-        &radiant::theme::ThemeTokens::default(),
-    );
-
-    assert!(frame_has_text(&frame, "Tag Editor"));
-    assert!(frame_has_text(&frame, "v Playback Type (2) [locked]"));
-    assert!(frame_has_text(&frame, "v Sound Type (2)"));
-    assert!(frame_has_text(&frame, "v Character (1)"));
-    assert!(frame_has_text(&frame, "v Prefix"));
-    assert!(frame_has_text(&frame, "v Tuning/Scale"));
-    assert!(frame_has_text(&frame, "loop"));
-    assert!(frame_has_text(&frame, "one-shot"));
-    assert!(frame_has_text(&frame, "hat"));
-    assert!(frame_has_text(&frame, "bass"));
-    assert!(frame_has_text(&frame, "seq"));
-}
-
-#[test]
-fn default_gui_tag_library_can_apply_default_playback_tags() {
-    let (mut state, _source_root, selected_file) = gui_state_with_temp_sample("tag-target.wav");
-
-    state.apply_message(
-        super::GuiMessage::ToggleMetadataTagLibrary,
-        &mut ui::UpdateContext::default(),
-    );
-    state.apply_message(
-        super::GuiMessage::ToggleMetadataTag(String::from("one-shot")),
-        &mut ui::UpdateContext::default(),
-    );
-
-    assert!(state.metadata_tag_library_open);
-    assert_eq!(
-        state.metadata_tags_by_file.get(&selected_file),
-        Some(&vec![String::from("one-shot")])
-    );
-}
-
-#[test]
-fn default_gui_tag_library_button_adds_existing_tag() {
-    let (mut state, _source_root, selected_file) = gui_state_with_temp_sample("tag-target.wav");
-    state.metadata_tags_by_file.insert(
-        String::from("other.wav"),
-        vec![String::from("bass"), String::from("hat")],
-    );
-
-    state.apply_message(
-        super::GuiMessage::ToggleMetadataTagLibrary,
-        &mut ui::UpdateContext::default(),
-    );
-    state.apply_message(
-        super::GuiMessage::ToggleMetadataTag(String::from("bass")),
-        &mut ui::UpdateContext::default(),
-    );
-
-    assert!(state.metadata_tag_library_open);
-    assert_eq!(
-        state.metadata_tags_by_file.get(&selected_file),
-        Some(&vec![String::from("bass")])
-    );
-}
-
-#[test]
-fn default_gui_tag_library_button_removes_selected_tag() {
-    let (mut state, _source_root, selected_file) = gui_state_with_temp_sample("tag-target.wav");
-    state.metadata_tags_by_file.insert(
-        selected_file.clone(),
-        vec![String::from("bass"), String::from("hat")],
-    );
-    state
-        .metadata_tags_by_file
-        .insert(String::from("other.wav"), vec![String::from("bass")]);
-
-    state.apply_message(
-        super::GuiMessage::ToggleMetadataTagLibrary,
-        &mut ui::UpdateContext::default(),
-    );
-    state.apply_message(
-        super::GuiMessage::ToggleMetadataTag(String::from("bass")),
-        &mut ui::UpdateContext::default(),
-    );
-
-    assert!(state.metadata_tag_library_open);
-    assert_eq!(
-        state.metadata_tags_by_file.get(&selected_file),
-        Some(&vec![String::from("hat")])
-    );
-    assert_eq!(state.sample_status, "Removed tag bass");
-}
-
-#[test]
-fn metadata_tag_chip_selection_can_be_deleted_from_selected_sample() {
-    let (mut state, _source_root, selected_file) = gui_state_with_temp_sample("tag-target.wav");
-    state.metadata_tags_by_file.insert(
-        selected_file.clone(),
-        vec![String::from("bass"), String::from("hat")],
-    );
-
-    state.apply_message(
-        super::GuiMessage::SelectMetadataTag(String::from("bass")),
-        &mut ui::UpdateContext::default(),
-    );
-    assert_eq!(state.selected_metadata_tag.as_deref(), Some("bass"));
-
-    state.apply_message(
-        super::GuiMessage::DeleteSelectedMetadataTag,
-        &mut ui::UpdateContext::default(),
-    );
-
-    assert_eq!(
-        state.metadata_tags_by_file.get(&selected_file),
-        Some(&vec![String::from("hat")])
-    );
-    assert_eq!(state.selected_metadata_tag, None);
-    assert_eq!(state.sample_status, "Removed tag bass");
-}
-
-#[test]
-fn default_gui_tag_library_category_headers_collapse_groups() {
-    let (mut state, _source_root, selected_file) = gui_state_with_temp_sample("tag-target.wav");
-    state
-        .metadata_tags_by_file
-        .insert(selected_file, vec![String::from("hat")]);
-    state.metadata_tag_library_open = true;
-
-    state.apply_message(
-        super::GuiMessage::ToggleMetadataTagCategory(String::from("sound-type")),
-        &mut ui::UpdateContext::default(),
-    );
-
-    let frame = radiant::runtime::UiSurface::new(super::view(&mut state).into_node()).frame(
-        Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(900.0, 620.0)),
-        &radiant::theme::ThemeTokens::default(),
-    );
-
-    assert!(frame_has_text(&frame, "> Sound Type (1)"));
-    assert!(!frame_has_text_after_x(&frame, "hat", DEFAULT_FOLDER_WIDTH));
-}
-
-#[test]
-fn default_gui_tag_library_drag_moves_tag_between_categories() {
-    let (mut state, _source_root, _selected_file) = gui_state_with_temp_sample("tag-target.wav");
-    state
-        .metadata_tags_by_file
-        .insert(String::from("other.wav"), vec![String::from("bass")]);
-
-    state.drag_metadata_tag(
-        String::from("bass"),
-        DragHandleMessage::Started {
-            position: Point::new(10.0, 10.0),
-        },
-        &mut ui::UpdateContext::default(),
-    );
-    state.drop_metadata_tag_on_category(
-        String::from("character"),
-        &mut ui::UpdateContext::default(),
-    );
-
-    assert_eq!(
-        state
-            .metadata_tag_dictionary
-            .get("bass")
-            .map(String::as_str),
-        Some("character")
-    );
-    assert_eq!(state.sample_status, "Moved tag bass to Character");
-}
-
-#[test]
-fn default_gui_tag_library_rejects_dragging_locked_playback_tags() {
-    let (mut state, _source_root, _selected_file) = gui_state_with_temp_sample("tag-target.wav");
-
-    state.drag_metadata_tag(
-        String::from("one-shot"),
-        DragHandleMessage::Started {
-            position: Point::new(10.0, 10.0),
-        },
-        &mut ui::UpdateContext::default(),
-    );
-
-    assert_eq!(state.metadata_tag_drag, None);
-    assert_eq!(state.metadata_tag_dictionary.get("one-shot"), None);
-    assert_eq!(state.sample_status, "Playback Type tags are locked");
-}
-
-#[test]
-fn default_gui_tag_library_pointer_drag_drops_tag_on_category_header() {
-    let (mut state, _source_root, _selected_file) = gui_state_with_temp_sample("tag-target.wav");
-    state
-        .metadata_tags_by_file
-        .insert(String::from("other.wav"), vec![String::from("bass")]);
-    state.metadata_tag_library_open = true;
-    let bridge = DeclarativeOwnedRuntimeBridge::new(
-        state,
-        |state| radiant::runtime::UiSurface::new(super::view(state).into_node()),
-        |state, message| state.apply_message(message, &mut ui::UpdateContext::default()),
-    );
-    let mut runtime = SurfaceRuntime::new(bridge, Vector2::new(900.0, 620.0));
-    let theme = radiant::theme::ThemeTokens::default();
-    let frame = runtime.frame(&theme);
-    let bass_rect = text_rect(&frame, "bass").expect("bass tag should paint");
-    let character_rect = text_rect(&frame, "v Character").expect("character header should paint");
-    let bass_point = Point::new(
-        (bass_rect.min.x + bass_rect.max.x) * 0.5,
-        (bass_rect.min.y + bass_rect.max.y) * 0.5,
-    );
-    let character_point = Point::new(
-        (character_rect.min.x + character_rect.max.x) * 0.5,
-        (character_rect.min.y + character_rect.max.y) * 0.5,
-    );
-
-    runtime.dispatch_event(Event::PointerPress {
-        position: bass_point,
-        button: PointerButton::Primary,
-        modifiers: PointerModifiers::default(),
-    });
-    runtime.dispatch_event(Event::PointerMove {
-        position: Point::new(bass_point.x + 8.0, bass_point.y + 2.0),
-    });
-    runtime.dispatch_event(Event::PointerRelease {
-        position: character_point,
-        button: PointerButton::Primary,
-        modifiers: PointerModifiers::default(),
-    });
-
-    assert_eq!(
-        runtime
-            .bridge()
-            .state()
-            .metadata_tag_dictionary
-            .get("bass")
-            .map(String::as_str),
-        Some("character")
-    );
-}
-
-#[test]
-fn default_gui_tag_library_right_click_opens_tag_context_menu() {
-    let (mut state, _source_root, _selected_file) = gui_state_with_temp_sample("tag-target.wav");
-    state
-        .metadata_tags_by_file
-        .insert(String::from("other.wav"), vec![String::from("oneshot")]);
-    state
-        .metadata_tag_dictionary
-        .insert(String::from("oneshot"), String::from("sound-type"));
-    state.metadata_tag_library_open = true;
-    let bridge = DeclarativeOwnedRuntimeBridge::new(
-        state,
-        |state| radiant::runtime::UiSurface::new(super::view(state).into_node()),
-        |state, message| state.apply_message(message, &mut ui::UpdateContext::default()),
-    );
-    let mut runtime = SurfaceRuntime::new(bridge, Vector2::new(900.0, 620.0));
-    let frame = runtime.frame(&radiant::theme::ThemeTokens::default());
-    let tag_rect = text_rect(&frame, "oneshot").expect("oneshot tag should paint");
-    let point = Point::new(
-        (tag_rect.min.x + tag_rect.max.x) * 0.5,
-        (tag_rect.min.y + tag_rect.max.y) * 0.5,
-    );
-
-    runtime.dispatch_event(Event::PointerPress {
-        position: point,
-        button: PointerButton::Secondary,
-        modifiers: PointerModifiers::default(),
-    });
-
-    let menu = runtime
-        .bridge()
-        .state()
-        .context_menu
-        .as_ref()
-        .expect("right-click should open metadata tag context menu");
-    assert_eq!(menu.kind, super::BrowserContextTargetKind::MetadataTag);
-    assert_eq!(menu.metadata_tag.as_deref(), Some("oneshot"));
-}
-
-#[test]
-fn metadata_tag_context_delete_removes_unlocked_global_tag() {
-    let (mut state, _source_root, selected_file) = gui_state_with_temp_sample("tag-target.wav");
-    state.metadata_tags_by_file.insert(
-        selected_file.clone(),
-        vec![String::from("oneshot"), String::from("hat")],
-    );
-    state.metadata_tags_by_file.insert(
-        String::from("other.wav"),
-        vec![String::from("oneshot"), String::from("seq")],
-    );
-    state
-        .metadata_tag_dictionary
-        .insert(String::from("oneshot"), String::from("sound-type"));
-    state.context_menu = Some(super::BrowserContextMenu {
-        kind: super::BrowserContextTargetKind::MetadataTag,
-        path: PathBuf::new(),
-        source_id: None,
-        metadata_tag: Some(String::from("oneshot")),
-        anchor: Point::new(12.0, 24.0),
-        title: String::from("oneshot"),
-    });
-
-    state.delete_context_metadata_tag(&mut ui::UpdateContext::default());
-
-    assert!(!state.metadata_tag_dictionary.contains_key("oneshot"));
-    assert_eq!(
-        state.metadata_tags_by_file.get(&selected_file),
-        Some(&vec![String::from("hat")])
-    );
-    assert_eq!(
-        state.metadata_tags_by_file.get("other.wav"),
-        Some(&vec![String::from("seq")])
-    );
-    assert_eq!(state.context_menu, None);
-    assert_eq!(
-        state.sample_status,
-        "Deleted tag oneshot from 2 assignment(s)"
-    );
-}
-
-#[test]
-fn metadata_tag_context_delete_rejects_locked_playback_tags() {
-    let (mut state, _source_root, _selected_file) = gui_state_with_temp_sample("tag-target.wav");
-    state.context_menu = Some(super::BrowserContextMenu {
-        kind: super::BrowserContextTargetKind::MetadataTag,
-        path: PathBuf::new(),
-        source_id: None,
-        metadata_tag: Some(String::from("loop")),
-        anchor: Point::new(12.0, 24.0),
-        title: String::from("loop"),
-    });
-
-    state.delete_context_metadata_tag(&mut ui::UpdateContext::default());
-
-    assert_eq!(state.sample_status, "Playback Type tags are locked");
-}
-
-#[test]
-fn default_gui_tag_library_uses_custom_dictionary_categories() {
-    let (mut state, _source_root, selected_file) = gui_state_with_temp_sample("tag-target.wav");
-    state
-        .metadata_tags_by_file
-        .insert(selected_file, vec![String::from("deep-kick")]);
-    state
-        .metadata_tag_dictionary
-        .insert(String::from("deep-kick"), String::from("sound-type"));
-    state.metadata_tag_library_open = true;
-
-    let frame = radiant::runtime::UiSurface::new(super::view(&mut state).into_node()).frame(
-        Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(900.0, 620.0)),
-        &radiant::theme::ThemeTokens::default(),
-    );
-
-    assert!(frame_has_text(&frame, "v Sound Type (1)"));
-    assert!(frame_has_text(&frame, "deep-kick"));
-    assert!(frame_has_text(&frame, "v Character"));
-    assert!(!frame_has_text(&frame, "v Character (1)"));
-}
-
-#[test]
-fn folder_browser_metadata_hides_tag_entry_when_no_file_is_selected() {
-    let browser = super::FolderBrowserState::load_default();
-    let tags = vec![String::from("kick")];
-    let frame = radiant::runtime::UiSurface::new(
-        super::folder_browser::folder_browser_view(
-            &browser,
-            260.0,
-            false,
-            "",
-            &[],
-            None,
-            "add tag",
-            None,
-            &[],
-            &tags,
-            &[],
-            None,
-        )
-        .into_node(),
-    )
-    .frame(
-        Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(260.0, 620.0)),
-        &radiant::theme::ThemeTokens::default(),
-    );
-
-    assert!(!frame_has_text(&frame, "Metadata"));
-    assert!(!frame_has_text(&frame, "Tags (1)"));
-    assert!(!frame_has_text(&frame, "kick"));
-    assert!(
-        !frame
-            .paint_plan
-            .primitives
-            .iter()
-            .any(|primitive| matches!(primitive, PaintPrimitive::TextInput(_)))
-    );
-}
-
-#[test]
-fn folder_browser_metadata_tags_grow_combined_entry_field() {
-    let browser = super::FolderBrowserState::load_default();
-    let small_tags = vec![String::from("kick")];
-    let larger_tags = vec![
-        String::from("kick"),
-        String::from("warm"),
-        String::from("one-shot"),
-        String::from("distorted"),
-    ];
-    let small = radiant::runtime::UiSurface::new(
-        super::folder_browser::folder_browser_view(
-            &browser,
-            260.0,
-            true,
-            "",
-            &[],
-            None,
-            "add tag",
-            None,
-            &[],
-            &small_tags,
-            &[],
-            None,
-        )
-        .into_node(),
-    )
-    .frame(
-        Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(260.0, 620.0)),
-        &radiant::theme::ThemeTokens::default(),
-    );
-    let larger = radiant::runtime::UiSurface::new(
-        super::folder_browser::folder_browser_view(
-            &browser,
-            260.0,
-            true,
-            "",
-            &[],
-            None,
-            "add tag",
-            None,
-            &[],
-            &larger_tags,
-            &[],
-            None,
-        )
-        .into_node(),
-    )
-    .frame(
-        Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(260.0, 620.0)),
-        &radiant::theme::ThemeTokens::default(),
-    );
-
-    assert!(frame_has_text(&larger, "distorted"));
-    assert!(!frame_has_text(&larger, "More"));
-    assert!(frame_has_clip_height(&small, 24.0));
-    let first_tag = text_rect(&larger, "kick").expect("first tag should paint");
-    let wrapped_tag = text_rect(&larger, "distorted").expect("wrapped tag should paint");
-    assert!(wrapped_tag.min.y > first_tag.min.y);
-}
-
-#[test]
-fn folder_browser_metadata_tag_field_caps_at_six_rows_then_scrolls() {
-    let browser = super::FolderBrowserState::load_default();
-    let tags = (0..24)
-        .map(|index| format!("tag-{index:02}"))
-        .collect::<Vec<_>>();
-    let frame = radiant::runtime::UiSurface::new(
-        super::folder_browser::folder_browser_view(
-            &browser,
-            260.0,
-            true,
-            "",
-            &[],
-            None,
-            "add tag",
-            None,
-            &[],
-            &tags,
-            &[],
-            None,
-        )
-        .into_node(),
-    )
-    .frame(
-        Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(260.0, 620.0)),
-        &radiant::theme::ThemeTokens::default(),
-    );
-
-    let tag_clip = frame.paint_plan.primitives.iter().find_map(|primitive| {
-        if let PaintPrimitive::ClipStart(clip) = primitive
-            && (clip.rect.height() - 129.0).abs() < 0.01
-        {
-            return Some(clip.rect);
-        }
-        None
-    });
-    assert!(
-        tag_clip.is_some(),
-        "combined tag field should clip overflowing tag rows"
-    );
-}
-
-#[test]
-fn metadata_tag_input_prompts_for_category_before_adding_new_tag() {
-    let config_base = tempfile::tempdir().expect("config base");
-    let _base_guard = wavecrate::app_dirs::ConfigBaseGuard::set(config_base.path().to_path_buf());
-    let (mut state, _source_root, selected_file) = gui_state_with_temp_sample("tag-target.wav");
-
-    state.apply_message(
-        super::GuiMessage::MetadataTagInput(radiant::widgets::TextInputMessage::Submitted {
-            value: String::from("Deep Kick"),
-        }),
-        &mut ui::UpdateContext::default(),
-    );
-
-    assert_eq!(state.metadata_tags_by_file.get(&selected_file), None);
-    assert_eq!(state.pending_metadata_tag_category_tag(), Some("deep-kick"));
-    assert_eq!(
-        state.metadata_tag_input_placeholder(),
-        "select group/parent tag"
-    );
-    assert_eq!(state.sample_status, "Choose a category for deep-kick");
-
-    state.apply_message(
-        super::GuiMessage::MetadataTagInput(radiant::widgets::TextInputMessage::Changed {
-            value: String::from("sound"),
-        }),
-        &mut ui::UpdateContext::default(),
-    );
-    assert_eq!(
-        state
-            .metadata_tag_completion_options()
-            .iter()
-            .find(|option| option.selected)
-            .map(|option| option.tag.as_str()),
-        Some("Sound Type")
-    );
-
-    state.apply_message(
-        super::GuiMessage::MetadataTagInput(radiant::widgets::TextInputMessage::Submitted {
-            value: String::from("sound"),
-        }),
-        &mut ui::UpdateContext::default(),
-    );
-
-    assert_eq!(
-        state.metadata_tags_by_file.get(&selected_file),
-        Some(&vec![String::from("deep-kick")])
-    );
-    assert_eq!(
-        state
-            .metadata_tag_dictionary
-            .get("deep-kick")
-            .map(String::as_str),
-        Some("sound-type")
-    );
-    assert_eq!(state.pending_metadata_tag_category_tag(), None);
-    assert!(state.metadata_tag_draft.is_empty());
-    assert_eq!(state.sample_status, "Added tag deep-kick");
-}
-
-#[test]
-fn metadata_tag_input_persists_tag_assignments_and_removals_to_source_database() {
-    let config_base = tempfile::tempdir().expect("config base");
-    let _base_guard = wavecrate::app_dirs::ConfigBaseGuard::set(config_base.path().to_path_buf());
-    let source_root = tempfile::tempdir().expect("source root");
-    let sample_path = source_root.path().join("persistent-tag.wav");
-    fs::write(&sample_path, []).expect("sample file");
-    wavecrate::sample_sources::config::save(&super::AppConfig {
-        sources: vec![wavecrate::sample_sources::SampleSource::new(
-            source_root.path().to_path_buf(),
-        )],
-        core: super::AppSettingsCore::default(),
-    })
-    .expect("seed config");
-    let selected_file = sample_path.display().to_string();
-    let mut state = GuiAppState::load_default().expect("default state loads");
-    state.folder_browser.select_file(selected_file.clone());
-
-    state.apply_message(
-        super::GuiMessage::MetadataTagInput(radiant::widgets::TextInputMessage::Submitted {
-            value: String::from("Deep Kick, Warm Tone"),
-        }),
-        &mut ui::UpdateContext::default(),
-    );
-    assert_eq!(
-        state.metadata_tags_by_file.get(&selected_file),
-        Some(&vec![String::from("deep-kick"), String::from("warm-tone")])
-    );
-
-    super::metadata_tags::persist_metadata_tag_additions_for_tests(
-        sample_path.clone(),
-        source_root.path().to_path_buf(),
-        PathBuf::from("persistent-tag.wav"),
-        vec![String::from("deep-kick"), String::from("warm-tone")],
-    )
-    .expect("persist tags");
-
-    let db = wavecrate::sample_sources::SourceDatabase::open(source_root.path())
-        .expect("open source db");
-    assert_eq!(
-        db.tag_labels_for_path(std::path::Path::new("persistent-tag.wav"))
-            .expect("tag labels"),
-        vec![String::from("deep-kick"), String::from("warm-tone")]
-    );
-
-    super::metadata_tags::persist_metadata_tag_removals_for_tests(
-        sample_path.clone(),
-        source_root.path().to_path_buf(),
-        PathBuf::from("persistent-tag.wav"),
-        vec![String::from("deep-kick")],
-    )
-    .expect("persist tag removal");
-
-    assert_eq!(
-        db.tag_labels_for_path(std::path::Path::new("persistent-tag.wav"))
-            .expect("tag labels after removal"),
-        vec![String::from("warm-tone")]
-    );
-
-    let reloaded = GuiAppState::load_default().expect("default state reloads");
-    assert_eq!(
-        reloaded.metadata_tags_by_file.get(&selected_file),
-        Some(&vec![String::from("warm-tone")])
-    );
-}
-
-#[test]
-fn metadata_tag_input_keeps_delimiters_while_editing() {
-    let mut state = gui_state_for_span_tests();
-
-    state.apply_message(
-        super::GuiMessage::MetadataTagInput(radiant::widgets::TextInputMessage::Changed {
-            value: String::from("kick, warm tone"),
-        }),
-        &mut ui::UpdateContext::default(),
-    );
-
-    assert!(state.metadata_tags_by_file.is_empty());
-    assert_eq!(state.metadata_tag_draft, "kick, warm tone");
-}
-
-#[test]
-fn metadata_tag_input_enters_selected_known_prefix() {
-    let (mut state, _source_root, selected_file) = gui_state_with_temp_sample("tag-target.wav");
-    state.metadata_tags_by_file.insert(
-        String::from("known-file"),
-        vec![String::from("kick"), String::from("warm")],
-    );
-
-    state.apply_message(
-        super::GuiMessage::MetadataTagInput(radiant::widgets::TextInputMessage::Changed {
-            value: String::from("ki"),
-        }),
-        &mut ui::UpdateContext::default(),
-    );
-    assert_eq!(
-        state.metadata_tag_completion_suffix().as_deref(),
-        Some("ck")
-    );
-
-    state.apply_message(
-        super::GuiMessage::MetadataTagInput(radiant::widgets::TextInputMessage::Submitted {
-            value: String::from("ki"),
-        }),
-        &mut ui::UpdateContext::default(),
-    );
-
-    assert_eq!(
-        state.metadata_tags_by_file.get(&selected_file),
-        Some(&vec![String::from("kick")])
-    );
-}
-
-#[test]
-fn metadata_tag_input_arrows_through_multiple_known_prefix_matches() {
-    let (mut state, _source_root, selected_file) = gui_state_with_temp_sample("tag-target.wav");
-    state.metadata_tags_by_file.insert(
-        String::from("known-file"),
-        vec![
-            String::from("kick"),
-            String::from("kicker"),
-            String::from("kind"),
-        ],
-    );
-
-    state.apply_message(
-        super::GuiMessage::MetadataTagInput(radiant::widgets::TextInputMessage::Changed {
-            value: String::from("ki"),
-        }),
-        &mut ui::UpdateContext::default(),
-    );
-    assert_eq!(
-        state
-            .metadata_tag_completion_options()
-            .iter()
-            .find(|option| option.selected)
-            .map(|option| option.tag.as_str()),
-        Some("kick")
-    );
-
-    state.apply_message(
-        super::GuiMessage::MoveMetadataTagCompletion(1),
-        &mut ui::UpdateContext::default(),
-    );
-    assert_eq!(
-        state
-            .metadata_tag_completion_options()
-            .iter()
-            .find(|option| option.selected)
-            .map(|option| option.tag.as_str()),
-        Some("kicker")
-    );
-
-    state.apply_message(
-        super::GuiMessage::MoveMetadataTagCompletion(1),
-        &mut ui::UpdateContext::default(),
-    );
-    assert_eq!(
-        state
-            .metadata_tag_completion_options()
-            .iter()
-            .find(|option| option.selected)
-            .map(|option| option.tag.as_str()),
-        Some("kind")
-    );
-
-    state.apply_message(
-        super::GuiMessage::MetadataTagInput(radiant::widgets::TextInputMessage::Submitted {
-            value: String::from("ki"),
-        }),
-        &mut ui::UpdateContext::default(),
-    );
-
-    assert_eq!(
-        state.metadata_tags_by_file.get(&selected_file),
-        Some(&vec![String::from("kind")])
-    );
-    assert!(state.metadata_tag_draft.is_empty());
-}
-
-#[test]
-fn folder_browser_metadata_tag_field_renders_completion_suffix_and_options() {
-    let browser = super::FolderBrowserState::load_default();
-    let completion_options = vec![
-        super::metadata_tags::MetadataTagCompletionOption {
-            tag: String::from("kick"),
-            category: "Sound Type",
-            selected: true,
-        },
-        super::metadata_tags::MetadataTagCompletionOption {
-            tag: String::from("kicker"),
-            category: "Character",
-            selected: false,
-        },
-    ];
-    let frame = radiant::runtime::UiSurface::new(
-        super::folder_browser::folder_browser_view(
-            &browser,
-            260.0,
-            true,
-            "ki",
-            &[String::from("kick")],
-            None,
-            "add tag",
-            Some("ck"),
-            completion_options.as_slice(),
-            &[String::from("warm")],
-            &[],
-            None,
-        )
-        .into_node(),
-    )
-    .frame(
-        Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(260.0, 620.0)),
-        &radiant::theme::ThemeTokens::default(),
-    );
-
-    assert!(frame_has_text(&frame, "kick"));
-    let tag_input = frame
-        .paint_plan
-        .primitives
-        .iter()
-        .find_map(|primitive| match primitive {
-            PaintPrimitive::TextInput(input) => Some(input),
-            _ => None,
-        })
-        .expect("tag input should paint");
-    assert_eq!(tag_input.state.value, "ki");
-    assert_eq!(tag_input.state.selection_anchor, 2);
-    assert_eq!(tag_input.state.caret, 2);
-    assert!(frame.paint_plan.primitives.iter().any(|primitive| {
-        matches!(primitive, PaintPrimitive::Text(text) if text.text.as_str() == "ck")
-    }));
-    assert!(frame_has_text(&frame, "Sound Type"));
-    assert!(frame_has_text(&frame, "kicker"));
-    assert!(frame_has_text(&frame, "Character"));
-    assert!(!frame_has_text(&frame, "Tab kick"));
-    assert!(frame_has_text(&frame, "warm"));
-    assert!(frame.paint_plan.primitives.iter().any(|primitive| {
-        matches!(
-            primitive,
-            PaintPrimitive::TextInput(input) if input.rect.height() <= 14.0
-        )
-    }));
-    assert!(frame.paint_plan.primitives.iter().any(|primitive| {
-        matches!(
-            primitive,
-            PaintPrimitive::FillRect(fill) if (fill.rect.height() - 18.0).abs() < 0.01
-        )
-    }));
-}
-
-#[test]
-fn metadata_autocomplete_suffix_is_not_editable_input_text() {
-    let (mut state, _source_root, _selected_file) = gui_state_with_temp_sample("tag-target.wav");
-    state
-        .metadata_tags_by_file
-        .insert(String::from("known.wav"), vec![String::from("kick")]);
-    state.metadata_tag_draft = String::from("ki");
-
-    let bridge = DeclarativeOwnedRuntimeBridge::new(
-        state,
-        |state| radiant::runtime::UiSurface::new(super::view(state).into_node()),
-        |state, message| state.apply_message(message, &mut ui::UpdateContext::default()),
-    );
-    let mut runtime = SurfaceRuntime::new(bridge, Vector2::new(900.0, 620.0));
-    let input_id = runtime
-        .frame(&radiant::theme::ThemeTokens::default())
-        .paint_plan
-        .primitives
-        .iter()
-        .find_map(|primitive| match primitive {
-            PaintPrimitive::TextInput(input) => Some(input.widget_id),
-            _ => None,
-        })
-        .expect("metadata tag input should paint");
-    assert!(runtime.focus_widget(input_id));
-
-    assert_eq!(
-        runtime.dispatch_focused_input(WidgetInput::KeyPress(WidgetKey::Backspace)),
-        Some(input_id)
-    );
-    assert_eq!(runtime.bridge().state().metadata_tag_draft, "k");
-    assert_eq!(
-        runtime.dispatch_focused_input(WidgetInput::KeyPress(WidgetKey::Backspace)),
-        Some(input_id)
-    );
-    assert!(runtime.bridge().state().metadata_tag_draft.is_empty());
-    assert!(!runtime.bridge().state().metadata_tag_completion_active());
-
-    let frame = runtime.frame(&radiant::theme::ThemeTokens::default());
-    let tag_input = frame
-        .paint_plan
-        .primitives
-        .iter()
-        .find_map(|primitive| match primitive {
-            PaintPrimitive::TextInput(input) if input.widget_id == input_id => Some(input),
-            _ => None,
-        })
-        .expect("metadata tag input should still paint");
-    assert!(tag_input.state.value.is_empty());
-    assert_eq!(tag_input.state.caret, 0);
-    assert_eq!(tag_input.state.selection_anchor, 0);
-    assert!(!frame.paint_plan.primitives.iter().any(
-        |primitive| matches!(primitive, PaintPrimitive::Text(text) if text.text.as_str() == "ick")
-    ));
-}
-
-#[test]
-fn metadata_autocomplete_does_not_block_sidebar_button_clicks() {
-    let (mut state, _source_root, _selected_file) = gui_state_with_temp_sample("tag-target.wav");
-    state
-        .metadata_tags_by_file
-        .insert(String::from("known.wav"), vec![String::from("kick")]);
-    state.metadata_tag_draft = String::from("ki");
-
-    let bridge = DeclarativeOwnedRuntimeBridge::new(
-        state,
-        |state| radiant::runtime::UiSurface::new(super::view(state).into_node()),
-        |state, message| state.apply_message(message, &mut ui::UpdateContext::default()),
-    );
-    let mut runtime = SurfaceRuntime::new(bridge, Vector2::new(900.0, 620.0));
-    let frame = runtime.frame(&radiant::theme::ThemeTokens::default());
-    let input_rect = frame
-        .paint_plan
-        .primitives
-        .iter()
-        .find_map(|primitive| match primitive {
-            PaintPrimitive::TextInput(input) => Some(input.rect),
-            _ => None,
-        })
-        .expect("metadata tag input should paint");
-    let input_point = Point::new(
-        (input_rect.min.x + input_rect.max.x) * 0.5,
-        (input_rect.min.y + input_rect.max.y) * 0.5,
-    );
-    runtime.dispatch_event(Event::PointerPress {
-        position: input_point,
-        button: PointerButton::Primary,
-        modifiers: PointerModifiers::default(),
-    });
-    runtime.dispatch_event(Event::PointerRelease {
-        position: input_point,
-        button: PointerButton::Primary,
-        modifiers: PointerModifiers::default(),
-    });
-    assert!(runtime.focused_widget().is_some());
-
-    let toggle_rect = text_rect(&runtime.frame(&radiant::theme::ThemeTokens::default()), ">")
-        .expect("tag library toggle should paint");
-    let point = Point::new(
-        (toggle_rect.min.x + toggle_rect.max.x) * 0.5,
-        (toggle_rect.min.y + toggle_rect.max.y) * 0.5,
-    );
-
-    runtime.dispatch_event(Event::PointerPress {
-        position: point,
-        button: PointerButton::Primary,
-        modifiers: PointerModifiers::default(),
-    });
-    runtime.dispatch_event(Event::PointerRelease {
-        position: point,
-        button: PointerButton::Primary,
-        modifiers: PointerModifiers::default(),
-    });
-
-    assert!(
-        runtime.bridge().state().metadata_tag_library_open,
-        "autocomplete popup must not prevent clicking the sidebar tag editor button"
-    );
-}
-
-#[test]
-fn metadata_autocomplete_does_not_block_folder_tree_clicks() {
-    let mut state = gui_state_for_span_tests();
-    let selected_file = state
-        .folder_browser
-        .selected_audio_files()
-        .first()
-        .expect("default browser should expose audio files")
-        .id
-        .clone();
-    state.folder_browser.select_file(selected_file);
-    state
-        .metadata_tags_by_file
-        .insert(String::from("known.wav"), vec![String::from("kick")]);
-    state.metadata_tag_draft = String::from("ki");
-    let selected_folder = state
-        .folder_browser
-        .selected_folder_path()
-        .expect("selected folder")
-        .display()
-        .to_string();
-
-    let bridge = DeclarativeOwnedRuntimeBridge::new(
-        state,
-        |state| radiant::runtime::UiSurface::new(super::view(state).into_node()),
-        |state, message| state.apply_message(message, &mut ui::UpdateContext::default()),
-    );
-    let mut runtime = SurfaceRuntime::new(bridge, Vector2::new(900.0, 620.0));
-    let frame = runtime.frame(&radiant::theme::ThemeTokens::default());
-    let input_rect = frame
-        .paint_plan
-        .primitives
-        .iter()
-        .find_map(|primitive| match primitive {
-            PaintPrimitive::TextInput(input) => Some(input.rect),
-            _ => None,
-        })
-        .expect("metadata tag input should paint");
-    let input_point = Point::new(
-        (input_rect.min.x + input_rect.max.x) * 0.5,
-        (input_rect.min.y + input_rect.max.y) * 0.5,
-    );
-    runtime.dispatch_event(Event::PointerPress {
-        position: input_point,
-        button: PointerButton::Primary,
-        modifiers: PointerModifiers::default(),
-    });
-    runtime.dispatch_event(Event::PointerRelease {
-        position: input_point,
-        button: PointerButton::Primary,
-        modifiers: PointerModifiers::default(),
-    });
-    assert!(runtime.focused_widget().is_some());
-
-    let frame = runtime.frame(&radiant::theme::ThemeTokens::default());
-    let (label, folder_rect) = frame
-        .paint_plan
-        .primitives
-        .iter()
-        .find_map(|primitive| match primitive {
-            PaintPrimitive::Text(text) if text.text.as_str().starts_with("[-] ") => {
-                Some((text.text.to_string(), text.rect))
-            }
-            _ => None,
-        })
-        .expect("expanded selected root folder should paint");
-    let point = Point::new(
-        (folder_rect.min.x + folder_rect.max.x) * 0.5,
-        (folder_rect.min.y + folder_rect.max.y) * 0.5,
-    );
-
-    runtime.dispatch_event(Event::PointerPress {
-        position: point,
-        button: PointerButton::Primary,
-        modifiers: PointerModifiers::default(),
-    });
-    runtime.dispatch_event(Event::PointerRelease {
-        position: point,
-        button: PointerButton::Primary,
-        modifiers: PointerModifiers::default(),
-    });
-
-    assert_eq!(
-        runtime
-            .bridge()
-            .state()
-            .folder_browser
-            .folder_expansion_for_tests(&selected_folder),
-        Some(false),
-        "autocomplete popup must not prevent clicking folder row {label}"
-    );
-}
-
-#[test]
-fn metadata_autocomplete_does_not_block_tag_library_clicks() {
-    let (mut state, _source_root, selected_file) = gui_state_with_temp_sample("tag-target.wav");
-    state.metadata_tags_by_file.insert(
-        String::from("known.wav"),
-        vec![String::from("kick"), String::from("bass")],
-    );
-    state.metadata_tag_draft = String::from("ki");
-    state.metadata_tag_library_open = true;
-
-    let bridge = DeclarativeOwnedRuntimeBridge::new(
-        state,
-        |state| radiant::runtime::UiSurface::new(super::view(state).into_node()),
-        |state, message| state.apply_message(message, &mut ui::UpdateContext::default()),
-    );
-    let mut runtime = SurfaceRuntime::new(bridge, Vector2::new(900.0, 620.0));
-    let frame = runtime.frame(&radiant::theme::ThemeTokens::default());
-    let input_rect = frame
-        .paint_plan
-        .primitives
-        .iter()
-        .find_map(|primitive| match primitive {
-            PaintPrimitive::TextInput(input) => Some(input.rect),
-            _ => None,
-        })
-        .expect("metadata tag input should paint");
-    let input_point = Point::new(
-        (input_rect.min.x + input_rect.max.x) * 0.5,
-        (input_rect.min.y + input_rect.max.y) * 0.5,
-    );
-    runtime.dispatch_event(Event::PointerPress {
-        position: input_point,
-        button: PointerButton::Primary,
-        modifiers: PointerModifiers::default(),
-    });
-    runtime.dispatch_event(Event::PointerRelease {
-        position: input_point,
-        button: PointerButton::Primary,
-        modifiers: PointerModifiers::default(),
-    });
-    assert!(runtime.focused_widget().is_some());
-
-    let tag_rect = text_rect(
-        &runtime.frame(&radiant::theme::ThemeTokens::default()),
-        "[ ] bass",
-    )
-    .expect("available tag should paint");
-    let point = Point::new(
-        (tag_rect.min.x + tag_rect.max.x) * 0.5,
-        (tag_rect.min.y + tag_rect.max.y) * 0.5,
-    );
-
-    runtime.dispatch_event(Event::PointerPress {
-        position: point,
-        button: PointerButton::Primary,
-        modifiers: PointerModifiers::default(),
-    });
-    runtime.dispatch_event(Event::PointerRelease {
-        position: point,
-        button: PointerButton::Primary,
-        modifiers: PointerModifiers::default(),
-    });
-
-    assert_eq!(
-        runtime
-            .bridge()
-            .state()
-            .metadata_tags_by_file
-            .get(&selected_file),
-        Some(&vec![String::from("bass")]),
-        "autocomplete popup must not prevent clicking tags in the tag library"
-    );
-}
-
-#[test]
-fn metadata_autocomplete_does_not_block_source_row_clicks_with_tag_library_open() {
-    let source_base = tempfile::tempdir().expect("source base");
-    let first_root = source_base.path().join("Alpha Samples");
-    let second_root = source_base.path().join("Beta Samples");
-    fs::create_dir_all(&first_root).expect("first source");
-    fs::create_dir_all(&second_root).expect("second source");
-    fs::write(first_root.join("alpha.wav"), []).expect("first sample");
-    fs::write(second_root.join("beta.wav"), []).expect("second sample");
-
-    let mut state = gui_state_for_span_tests();
-    state.folder_browser = super::FolderBrowserState::from_sample_sources(&[
-        wavecrate::sample_sources::SampleSource::new(first_root.clone()),
-        wavecrate::sample_sources::SampleSource::new(second_root.clone()),
-    ]);
-    let first_file = first_root.join("alpha.wav").display().to_string();
-    state.folder_browser.select_file(first_file);
-    state
-        .metadata_tags_by_file
-        .insert(String::from("known.wav"), vec![String::from("kick")]);
-    state.metadata_tag_draft = String::from("ki");
-    state.metadata_tag_library_open = true;
-
-    let bridge = DeclarativeOwnedRuntimeBridge::new(
-        state,
-        |state| radiant::runtime::UiSurface::new(super::view(state).into_node()),
-        |state, message| state.apply_message(message, &mut ui::UpdateContext::default()),
-    );
-    let mut runtime = SurfaceRuntime::new(bridge, Vector2::new(589.0, 571.0));
-    let frame = runtime.frame(&radiant::theme::ThemeTokens::default());
-    let input_rect = frame
-        .paint_plan
-        .primitives
-        .iter()
-        .find_map(|primitive| match primitive {
-            PaintPrimitive::TextInput(input) => Some(input.rect),
-            _ => None,
-        })
-        .expect("metadata tag input should paint");
-    let input_point = Point::new(
-        (input_rect.min.x + input_rect.max.x) * 0.5,
-        (input_rect.min.y + input_rect.max.y) * 0.5,
-    );
-    runtime.dispatch_event(Event::PointerPress {
-        position: input_point,
-        button: PointerButton::Primary,
-        modifiers: PointerModifiers::default(),
-    });
-    runtime.dispatch_event(Event::PointerRelease {
-        position: input_point,
-        button: PointerButton::Primary,
-        modifiers: PointerModifiers::default(),
-    });
-    assert!(runtime.focused_widget().is_some());
-
-    let source_rect = text_rect(
-        &runtime.frame(&radiant::theme::ThemeTokens::default()),
-        "Beta Samples",
-    )
-    .expect("second source should paint");
-    let point = Point::new(
-        (source_rect.min.x + source_rect.max.x) * 0.5,
-        (source_rect.min.y + source_rect.max.y) * 0.5,
-    );
-    runtime.dispatch_event(Event::PointerPress {
-        position: point,
-        button: PointerButton::Primary,
-        modifiers: PointerModifiers::default(),
-    });
-    runtime.dispatch_event(Event::PointerRelease {
-        position: point,
-        button: PointerButton::Primary,
-        modifiers: PointerModifiers::default(),
-    });
-
-    assert_eq!(
-        runtime
-            .bridge()
-            .state()
-            .folder_browser
-            .selected_folder_path(),
-        Some(second_root),
-        "autocomplete popup and tag library must not prevent clicking source rows"
-    );
-}
-
-#[test]
-fn metadata_tag_field_background_click_focuses_tag_input() {
-    let (state, _source_root, _selected_file) = gui_state_with_temp_sample("tag-target.wav");
-    let bridge = DeclarativeOwnedRuntimeBridge::new(
-        state,
-        |state| radiant::runtime::UiSurface::new(super::view(state).into_node()),
-        |state, message| state.apply_message(message, &mut ui::UpdateContext::default()),
-    );
-    let mut runtime = SurfaceRuntime::new(bridge, Vector2::new(900.0, 620.0));
-    let frame = runtime.frame(&radiant::theme::ThemeTokens::default());
-    let input_rect = frame
-        .paint_plan
-        .primitives
-        .iter()
-        .find_map(|primitive| match primitive {
-            PaintPrimitive::TextInput(input) => Some(input.rect),
-            _ => None,
-        })
-        .expect("tag input should paint");
-    let input_id = text_input_widget_id(&frame).expect("tag input widget id");
-    assert!(
-        input_rect.width() > 160.0,
-        "tag input should expose a broad click target, got {input_rect:?}"
-    );
-    let point = Point::new(
-        input_rect.min.x + 10.0,
-        (input_rect.min.y + input_rect.max.y) * 0.5,
-    );
-
-    runtime.dispatch_event(Event::PointerPress {
-        position: point,
-        button: PointerButton::Primary,
-        modifiers: PointerModifiers::default(),
-    });
-    runtime.dispatch_event(Event::PointerRelease {
-        position: point,
-        button: PointerButton::Primary,
-        modifiers: PointerModifiers::default(),
-    });
-    let target = runtime.widget_at(point);
-    let target_rect = target.and_then(|widget_id| {
-        runtime
-            .frame(&radiant::theme::ThemeTokens::default())
-            .paint_plan
-            .first_widget_rect(widget_id)
-    });
-
-    assert_eq!(
-        runtime.focused_widget(),
-        Some(input_id),
-        "clicking the visible tag-entry field should focus the embedded tag input; target={target:?} target_rect={target_rect:?} input_rect={input_rect:?}"
-    );
-}
-
-#[test]
-fn folder_browser_metadata_tag_field_renders_pending_category_prompt() {
-    let browser = super::FolderBrowserState::load_default();
-    let completion_options = vec![super::metadata_tags::MetadataTagCompletionOption {
-        tag: String::from("Sound Type"),
-        category: "Group",
-        selected: true,
-    }];
-    let frame = radiant::runtime::UiSurface::new(
-        super::folder_browser::folder_browser_view(
-            &browser,
-            260.0,
-            true,
-            "sound",
-            &[],
-            Some("deep-kick"),
-            "select group/parent tag",
-            Some("-type"),
-            completion_options.as_slice(),
-            &[],
-            &[],
-            None,
-        )
-        .into_node(),
-    )
-    .frame(
-        Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(260.0, 620.0)),
-        &radiant::theme::ThemeTokens::default(),
-    );
-
-    assert!(frame_has_text(&frame, "deep-kick ->"));
-    assert!(frame_has_text(&frame, "Sound Type"));
-    assert!(frame_has_text(&frame, "Group"));
-    let pending_tag_rect = text_rect(&frame, "deep-kick ->").expect("pending tag should paint");
-    let category_input = frame
-        .paint_plan
-        .primitives
-        .iter()
-        .find_map(|primitive| match primitive {
-            PaintPrimitive::TextInput(input) => Some(input),
-            _ => None,
-        })
-        .expect("category input should paint");
-    assert_eq!(category_input.state.value, "sound");
-    assert_eq!(category_input.state.selection_anchor, 5);
-    assert_eq!(category_input.state.caret, 5);
-    assert!(frame.paint_plan.primitives.iter().any(|primitive| {
-        matches!(primitive, PaintPrimitive::Text(text) if text.text.as_str() == "-type")
-    }));
-    assert!(
-        category_input.rect.min.x > pending_tag_rect.max.x,
-        "category input should stay on the same row after the pending tag arrow"
-    );
-    let sound_type_rect = text_rect(&frame, "Sound Type").expect("completion option should paint");
-    assert!(
-        sound_type_rect.max.y < pending_tag_rect.min.y,
-        "completion popup should expand upward above the tag input"
-    );
-}
-
-#[test]
-fn folder_browser_metadata_tag_input_moves_to_next_row_when_crowded() {
-    let browser = super::FolderBrowserState::load_default();
-    let tags = vec![
-        String::from("test"),
-        String::from("another"),
-        String::from("cool-tag"),
-    ];
-    let frame = radiant::runtime::UiSurface::new(
-        super::folder_browser::folder_browser_view(
-            &browser,
-            260.0,
-            true,
-            "wow",
-            &[],
-            None,
-            "add tag",
-            None,
-            &[],
-            &tags,
-            &[],
-            None,
-        )
-        .into_node(),
-    )
-    .frame(
-        Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(260.0, 620.0)),
-        &radiant::theme::ThemeTokens::default(),
-    );
-
-    let first_tag_y = frame.paint_plan.primitives.iter().find_map(|primitive| {
-        if let PaintPrimitive::FillRect(fill) = primitive
-            && (fill.rect.height() - 18.0).abs() < 0.01
-        {
-            return Some(fill.rect.min.y);
-        }
-        None
-    });
-    let first_tag_y = first_tag_y.expect("tag pill should paint in the tag field");
-    let input_rect = frame.paint_plan.primitives.iter().find_map(|primitive| {
-        if let PaintPrimitive::TextInput(input) = primitive {
-            return Some(input.rect);
-        }
-        None
-    });
-    let input_rect = input_rect.expect("tag input should paint");
-
-    assert!(input_rect.min.y > first_tag_y);
-    assert!(input_rect.max.x <= 260.0);
-}
-
-#[test]
-fn folder_browser_metadata_tag_input_keeps_identity_when_wrapping_rows() {
-    let browser = super::FolderBrowserState::load_default();
-    let short_tags = vec![String::from("kick")];
-    let crowded_tags = vec![
-        String::from("test"),
-        String::from("another"),
-        String::from("cool-tag"),
-    ];
-    let short_frame = radiant::runtime::UiSurface::new(
-        super::folder_browser::folder_browser_view(
-            &browser,
-            260.0,
-            true,
-            "wow",
-            &[],
-            None,
-            "add tag",
-            None,
-            &[],
-            &short_tags,
-            &[],
-            None,
-        )
-        .into_node(),
-    )
-    .frame(
-        Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(260.0, 620.0)),
-        &radiant::theme::ThemeTokens::default(),
-    );
-    let crowded_frame = radiant::runtime::UiSurface::new(
-        super::folder_browser::folder_browser_view(
-            &browser,
-            260.0,
-            true,
-            "wow",
-            &[],
-            None,
-            "add tag",
-            None,
-            &[],
-            &crowded_tags,
-            &[],
-            None,
-        )
-        .into_node(),
-    )
-    .frame(
-        Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(260.0, 620.0)),
-        &radiant::theme::ThemeTokens::default(),
-    );
-
-    let short_input = text_input_widget_id(&short_frame).expect("short tag field input");
-    let crowded_input = text_input_widget_id(&crowded_frame).expect("crowded tag field input");
-
-    assert_eq!(short_input, crowded_input);
-}
-
-#[test]
-fn folder_browser_metadata_tag_input_wraps_after_full_tag_row() {
-    let browser = super::FolderBrowserState::load_default();
-    let tags = vec![
-        String::from("yay"),
-        String::from("cool-tag"),
-        String::from("thing"),
-        String::from("potato"),
-    ];
-    let frame = radiant::runtime::UiSurface::new(
-        super::folder_browser::folder_browser_view(
-            &browser,
-            450.0,
-            true,
-            "",
-            &[],
-            None,
-            "add tag",
-            None,
-            &[],
-            &tags,
-            &[],
-            None,
-        )
-        .into_node(),
-    )
-    .frame(
-        Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(450.0, 620.0)),
-        &radiant::theme::ThemeTokens::default(),
-    );
-
-    let first_tag_y = frame.paint_plan.primitives.iter().find_map(|primitive| {
-        if let PaintPrimitive::FillRect(fill) = primitive
-            && (fill.rect.height() - 18.0).abs() < 0.01
-        {
-            return Some(fill.rect.min.y);
-        }
-        None
-    });
-    let first_tag_y = first_tag_y.expect("tag pill should paint in the tag field");
-    let input_rect = frame.paint_plan.primitives.iter().find_map(|primitive| {
-        if let PaintPrimitive::TextInput(input) = primitive {
-            return Some(input.rect);
-        }
-        None
-    });
-    let input_rect = input_rect.expect("tag input should paint");
-
-    assert!(input_rect.min.y > first_tag_y);
-    assert!(input_rect.max.x <= 450.0);
 }
 
 #[test]
