+++ src/gui_app/tests/metadata_tag_tests.rs
@@ -1,0 +1,5 @@
+mod autocomplete;
+mod chips;
+mod input;
+mod layout;
+mod library;
